enable app router
