+++ package.json
@@ -30,7 +30,7 @@
     "dayjs": "^1.11.7",
     "eslint": "8.38.0",
     "eslint-config-next": "13.3.0",
-    "next": "13.3.0",
+    "next": "13.4.7",
     "oidc-client-ts": "^2.2.3",
     "postcss": "8.4.23",
     "react": "18.2.0",
